+++ src/operators/index.ts
@@ -1,3 +1,2 @@
 export * from './derive.operator';
-export * from './omit.operator';
 export * from './pick.operator';